Remove browserFS
