+++ packages/website/package.json
@@ -42,7 +42,6 @@
     "babel-eslint": "10.x",
     "babel-loader": "^8.0.6",
     "babel-plugin-istanbul": "^6.0.0",
-    "browserfs": "^2.0.0",
     "copy-webpack-plugin": "^5.1.1",
     "core-js": "3",
     "cypress": "^4.2.0",
@@ -55,14 +54,12 @@
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "7.x",
     "eslint-plugin-react-hooks": "2.x",
-    "fs": "^0.0.1-security",
     "http-server": "^0.12.1",
     "istanbul-lib-coverage": "^3.0.0",
     "kuromoji": "^0.1.2",
     "negaposi": "^1.2.0",
     "nyc": "^15.0.0",
     "path": "^0.12.7",
-    "raw-loader": "^4.0.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-helmet": "^6.0.0",
