+++ packages/website/resources/index.html
@@ -44,7 +44,6 @@
     <div class="app"></div>
 
     <script src="main.js"></script>
-    <script src="vendors~main.js"></script>
 
     <noscript>
       校正さんを使用するには、JavaScript を有効にしてください。
