--- packages/website/src/yml.d.ts
@@ -1,1 +1,0 @@
-declare module '*.yml';
